RaavareBatchResources in own class and removed them from stockresources
